requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.11.0
osqp>=0.6.3  # For QP-based MPC optimization
pyyaml>=6.0.1  # For config file handling
pytest>=8.0.0  # For testing
cvxopt>=1.3.0  # For QP solving
control>=0.9.0  # For control system analysis
casadi>=3.5.5  # For nonlinear optimization
